+++ src/main/java/com/example/demo/Repository/EmailKodRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmailKodRepository extends CrudRepository<EmailKod,Integer> {
+public interface EmailKodRepository extends CrudRepository<EmailKod,Long> {
     EmailKod findByKod(String kod);
     EmailKod findByEmail(String email);
 }
